src/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Adieva15",email = "adievadine@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot (>=22.5,<23.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "fastapi (>=0.122.0,<0.123.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "aiogram (>=3.22.0,<4.0.0)",
    "tortoise-orm (>=0.25.1,<0.26.0)",
    "aerich (>=0.9.2,<0.10.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "tomlkit (>=0.13.3,<0.14.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "config_reader.TORTOISE_ORM"
location = "./db/migrations"
src_folder = "./."
